pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["Margulanz <margulan.zholaman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
